requirements.txt
Pillow==8.3.1
pytesseract==0.3.8
googletrans==3.1.0a0
pyttsx3 2.90
SpeechRecognition 3.8.1
opencv-python 4.5.3.56